Refactor moving ReqMessage to utils/msgs
