+++ src/UltralyticsBot/utils/msgs.py
@@ -5,14 +5,15 @@ Date: 2023-10-29
 
 Requires: requests
 """
-import io
+import re
 from typing import Callable
 
+import discord
 import requests
-from discord import File
 
-from UltralyticsBot import GH
-from UltralyticsBot.utils.general import dec2str, align_boxcoord, is_link
+from UltralyticsBot import GH, BOT_ID
+from UltralyticsBot.utils.general import dec2str, align_boxcoord
+from UltralyticsBot.utils.checks import is_link, is_img_link, URL_RGX
 
 NEWLINE = '\n' # use with f-strings
 BOX_LJUST = 24 # Box coordinates will always be -> '(1234, 1234, 1234, 1234)'
@@ -70,10 +71,92 @@ class ResponseMsg():
             self.msg += gen_title(self.cls_pad) if self.txt else ''
             
             self.anno_im, self.result_txt = plt_fn(predictions=self.data)
-            self.msg += ((self.result_txt) if self.txt and self.result_txt != '' else '') + '```'
+            self.msg += ((self.result_txt + '```') if self.txt else ('```' if self.txt and self.result_txt != '' else ''))
         else:
             self.anno_im = None
             self.msg = API_ERR_MSG.format(self.code, self.reason)
         
         return (self.anno_im, self.msg)
 
+class ReqMessage:
+    """
+    Attributes
+    ---
+    msg - ``discord.Message``
+        Discord message for inference request.
+
+    url - ``str`` | ``None``
+        URL string when `self.msg` contains valid URL, otherwise ``None``.
+
+    author - ``discord.Message.author`` | ``None``
+        Discord message author.
+
+    mentions - ``list``
+        List of mentioned users in `self.msg`.
+
+    media - ``list[discord.Attachment]``
+        List of media attached to `self.msg`.
+
+    im_width - ``int`` | ``None``
+        Width in pixels of image attched to `self.msg` if any, otherwise ``None``.
+
+    im_height - ``int`` | ``None``
+        Height in pixels of image attched to `self.msg` if any, otherwise ``None``.
+
+    has_url - ``bool``
+        Is `True` when text from `self.msg` contains what appears to be valid URL string, otherwise `False`.
+
+    has_media - ``bool``
+        Is `True` when `self.msg` contains any attachments, otherwise `False`.
+
+    has_text - ``bool``
+        Is `True` when `self.msg` contains text other than triggering-keywords and whitespace, otherwise `False`.
+
+    has_img - ``bool``
+        Is `True` when `self.msg` has attachment with content type 'image', otherwise `False`
+
+    bot_mention - ``bool``
+        Is `True` when `self.msg` contains bot-user mention/tag, otherwise `False`.
+
+    Methods
+    ---
+    check_message() - Executes during `__init__` and populates attributes from `discord.Message`
+
+    get_url() - If attributes not populated, executes `check_message()` then returns first URL as ``str`` found in `discord.Message.content`, otherwise returns ``None``.
+    """
+    commands = ['predict', 'help', 'about', 'hub'] # NOTE maybe pull from YAML instead
+    
+    def __init__(self, msg:discord.Message) -> None:
+        self.msg = msg
+        self.url = self.author = self.mentions = self.media = None
+        self.im_height = self.im_width = self.attached_im = self.img_size = None
+        self.has_url = self.has_media = self.has_text = self.has_img = self.bot_mention = False
+        self.check_message()
+        
+    def check_message(self) -> None:
+        self.has_text = isinstance(self.msg.content,str) and any(self.msg.content.replace("$predict","").strip())
+        self.media = self.msg.attachments if any(self.msg.attachments) else []
+        
+        self.has_url = is_link(self.msg.content) if self.has_text else False
+        self.has_img = (['image' in a.content_type for a in self.media]) if any(self.media) else False
+        
+        self.author = self.msg.author
+        self.mentions = self.msg.mentions
+        self.bot_mention = BOT_ID in [m.id for m in self.mentions]
+        
+        if self.has_text and self.has_url:
+            self.url = re.search(URL_RGX, self.msg.content, re.IGNORECASE).group()
+        
+        elif self.has_img:
+            self.attached_im = [a for a in self.media if 'image' in a.content_type][0] # only allow one image
+            self.url = self.attached_im.url
+            self.im_height, self.im_width = self.attached_im.height, self.attached_im.width
+            self.img_size = self.attached_im.size / (1024 ** 2)
+    
+    def get_url(self) -> str:
+        _ = self.check_message() if self.url is None else None
+        return self.url
+    
+    def media_info(self) -> tuple[int|None,int|None,float|None]:
+        """If image was attached, returns image height, width, and file-size."""
+        return self.im_height, self.im_width, self.img_size
